gpu subsystem cleanup #2
