+++ crates/darkly/src/engine/painting.rs
@@ -693,7 +693,7 @@ impl DarklyEngine {
 
         // Refresh the blend-uniform buffer so the composite pass sees the
         // new offset/size on the next render.
-        self.compositor.update_layer_uniforms_full(
+        self.compositor.update_layer_uniforms(
             &self.gpu.queue,
             layer_id,
             opacity,
