+++ crates/darkly/src/gpu/compositor.rs
@@ -4,7 +4,7 @@ use crate::gpu::atlas::LayerTexture;
 use crate::gpu::blend::BlendPipelines;
 use crate::gpu::content_bounds::ContentBoundsPass;
 use crate::gpu::effect::EffectCache;
-use crate::gpu::overlay::{OverlayPrimitive, ToolOverlay};
+use crate::gpu::overlay::ToolOverlay;
 use crate::gpu::params::ParamValue;
 use crate::gpu::veil_chain::VeilChain;
 use crate::gpu::view::{ViewTransform, DEFAULT_WORKSPACE_BG};
@@ -118,19 +118,19 @@ struct GroupState {
 /// [`LayerContent`]. One pool keyed by [`LayerId`] replaces the previous
 /// `raster_cache` + `void_layers` split, so the compositor's lookup paths
 /// (blend arm, uniforms write, dispose) don't dispatch on layer kind.
-struct LayerCache {
+pub(super) struct LayerCache {
     /// Uniform buffer holding opacity + blend_mode + isolated.
-    uniform_buf: wgpu::Buffer,
+    pub(super) uniform_buf: wgpu::Buffer,
     /// CPU-side cache of the blend properties last written to `uniform_buf`.
     /// Kept here so the floating-preview path can mirror them into its own
     /// canvas-aligned uniform buffer without re-reading the GPU buffer.
-    opacity: f32,
+    pub(super) opacity: f32,
     /// Cached gpu_value for the layer's blend mode. The compositor never
     /// branches on which mode this is — the shader does — so we mirror the
     /// raw shader integer rather than carry a registration pointer through
     /// every per-frame access.
-    blend_mode: u32,
-    isolated: bool,
+    pub(super) blend_mode: u32,
+    pub(super) isolated: bool,
     /// Where this layer's pixels come from. Raster pixels arrive via paint;
     /// procedural pixels are regenerated on demand by a [`Void`] trait
     /// object before each composite.
@@ -173,18 +173,18 @@ struct ProceduralContent {
 /// translates per-pixel from canvas UV to layer UV.
 #[repr(C)]
 #[derive(Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct BlendUniforms {
-    opacity: f32,
-    blend_mode: u32,
-    isolated: u32,
-    _pad1: f32,
+pub(super) struct BlendUniforms {
+    pub(super) opacity: f32,
+    pub(super) blend_mode: u32,
+    pub(super) isolated: u32,
+    pub(super) _pad1: f32,
     /// Layer's (offset_x, offset_y) in canvas coordinates.
-    layer_offset: [f32; 2],
+    pub(super) layer_offset: [f32; 2],
     /// Layer texture dimensions in pixels.
-    layer_size: [f32; 2],
+    pub(super) layer_size: [f32; 2],
     /// Canvas dimensions in pixels.
-    canvas_size: [f32; 2],
-    _pad2: [f32; 2],
+    pub(super) canvas_size: [f32; 2],
+    pub(super) _pad2: [f32; 2],
 }
 
 /// GPU state for a passthrough group that has a mask (Photoshop-style).
@@ -214,7 +214,7 @@ pub struct Compositor {
     /// layer textures (Rgba8Unorm), mask modifier textures (R8Unorm), and
     /// any future pixel-bearing modifier kinds — `LayerTexture.format`
     /// distinguishes them. One lookup per access, no fan-out.
-    node_textures: HashMap<LayerId, LayerTexture>,
+    pub(super) node_textures: HashMap<LayerId, LayerTexture>,
 
     /// Default mask bind group using the 1×1 white texture (pass-through
     /// fallback for hosts without a visible mask modifier).
@@ -239,9 +239,9 @@ pub struct Compositor {
     /// the document's [`LayerId`] — both kinds share the same blend
     /// pipeline path, so collapsing them into one pool means the blend
     /// arm, uniforms write, and dispose all do exactly one lookup.
-    layer_cache: HashMap<LayerId, LayerCache>,
+    pub(super) layer_cache: HashMap<LayerId, LayerCache>,
 
-    blend_pipelines: BlendPipelines,
+    pub(super) blend_pipelines: BlendPipelines,
 
     // --- Passthrough Group Mask (Photoshop-style snapshot-lerp) ---
     mask_lerp_pipeline: wgpu::RenderPipeline,
@@ -257,7 +257,7 @@ pub struct Compositor {
     /// View transform uniform buffer for the present shader.
     view_uniform_buf: wgpu::Buffer,
 
-    sampler: wgpu::Sampler,
+    pub(super) sampler: wgpu::Sampler,
 
     /// Dirty gate — false means nothing changed, skip compositing.
     needs_composite: bool,
@@ -274,8 +274,8 @@ pub struct Compositor {
     /// uniformly.
     dirty_node_pixels: HashSet<LayerId>,
 
-    canvas_width: u32,
-    canvas_height: u32,
+    pub(super) canvas_width: u32,
+    pub(super) canvas_height: u32,
     /// Padded (tile-aligned) render target dimensions — used for shader UV
     /// computations in the transform pass, which must match the actual
     /// accumulator texture size.
@@ -290,7 +290,7 @@ pub struct Compositor {
     void_registry: VoidRegistry,
 
     // --- Floating Content Transform ---
-    transform_pass: crate::gpu::transform::TransformPass,
+    pub(super) transform_pass: crate::gpu::transform::TransformPass,
 
     // --- Isolation (session state) ---
     /// When `Some(id)`, the render walk descends only into nodes on the
@@ -1072,11 +1072,7 @@ impl Compositor {
         queue: &wgpu::Queue,
         node_id: LayerId,
     ) {
-        let Some(tex) = self
-            .node_textures
-            .get(&node_id)
-            .or_else(|| self.node_textures.get(&node_id))
-        else {
+        let Some(tex) = self.node_textures.get(&node_id) else {
             return;
         };
         let r_channel = tex.format() == wgpu::TextureFormat::R8Unorm;
@@ -1806,29 +1802,15 @@ impl Compositor {
     }
 
     /// Update a content layer's uniforms (called when opacity, blend mode,
-    /// or isolated changes). Convenience wrapper that pins `isolated` to
-    /// false — bulk sync paths that compute isolation already use the full
-    /// variant directly.
-    pub fn update_layer_uniforms(
-        &mut self,
-        queue: &wgpu::Queue,
-        layer_id: LayerId,
-        opacity: f32,
-        blend_mode_gpu: u32,
-    ) {
-        self.update_layer_uniforms_full(queue, layer_id, opacity, blend_mode_gpu, false);
-    }
-
-    /// Update a content layer's uniforms including the isolated flag. Works
-    /// uniformly for raster and procedural layers — both store their blend
-    /// state in the same [`LayerCache`] and sample from canvas-positioned
-    /// textures in `node_textures`. Reads the layer's bounds from its
-    /// `LayerTexture` so callers don't need to thread them through;
-    /// bounds-changing operations update the texture's stored offset/size
-    /// directly via `resize_node_texture`.
+    /// or isolated changes). Works uniformly for raster and procedural
+    /// layers — both store their blend state in the same [`LayerCache`]
+    /// and sample from canvas-positioned textures in `node_textures`.
+    /// Reads the layer's bounds from its `LayerTexture` so callers don't
+    /// need to thread them through; bounds-changing operations update the
+    /// texture's stored offset/size directly via `resize_node_texture`.
     ///
     /// `blend_mode_gpu` is the registry-resolved gpu_value.
-    pub fn update_layer_uniforms_full(
+    pub fn update_layer_uniforms(
         &mut self,
         queue: &wgpu::Queue,
         layer_id: LayerId,
@@ -1869,39 +1851,6 @@ impl Compositor {
         self.write_preview_blend_uniforms_if_active(queue, layer_id);
     }
 
-    /// Write the floating preview's canvas-aligned blend uniforms using the
-    /// given layer's cached blend props. No-op when there is no active
-    /// floating, or when the active floating's target is not `layer_id`.
-    /// Called from both `update_layer_uniforms_full` (on prop change) and
-    /// the floating setup paths (to seed the buffer at session start).
-    fn write_preview_blend_uniforms_if_active(&self, queue: &wgpu::Queue, layer_id: LayerId) {
-        let state = match self.transform_pass.active.as_ref() {
-            Some(s) if s.target_layer == layer_id => s,
-            _ => return,
-        };
-        let cache = match self.layer_cache.get(&layer_id) {
-            Some(c) => c,
-            None => return,
-        };
-        let uniforms = BlendUniforms {
-            opacity: cache.opacity,
-            blend_mode: cache.blend_mode,
-            isolated: cache.isolated as u32,
-            _pad1: 0.0,
-            layer_offset: [0.0, 0.0],
-            layer_size: [self.canvas_width as f32, self.canvas_height as f32],
-            canvas_size: [self.canvas_width as f32, self.canvas_height as f32],
-            _pad2: [0.0, 0.0],
-        };
-        queue.write_buffer(
-            &state.preview_blend_uniform_buf,
-            0,
-            bytemuck::bytes_of(&uniforms),
-        );
-    }
-
-    /// Look up the resolved mask bind group for a modifier id, falling back to
-    /// the default (1x1 white = no masking) when no real mask is active.
     /// Get the composited output texture (root group's composite cache).
     /// Used by the color picker for readback.
     pub fn composited_texture(&self) -> &wgpu::Texture {
@@ -1920,54 +1869,16 @@ impl Compositor {
         &mut self.veil_chain
     }
 
-    // --- Tool Overlay ---
-
-    /// Replace the current overlay primitives.
-    pub fn set_overlay_primitives(&mut self, prims: Vec<OverlayPrimitive>) {
-        self.tool_overlay.set_primitives(prims);
-        self.needs_present = true;
+    /// Read-only access to the tool overlay. Callers do their own dispatch;
+    /// the compositor stops being a switchboard.
+    pub fn tool_overlay(&self) -> &ToolOverlay {
+        &self.tool_overlay
     }
 
-    /// Clear all overlay primitives.
-    pub fn clear_overlay(&mut self) {
-        self.tool_overlay.clear_primitives();
-        self.needs_present = true;
-    }
-
-    /// Advance overlay animation time.
-    pub fn update_overlay_time(&mut self, dt: f32) {
-        self.tool_overlay.advance_time(dt);
-    }
-
-    /// Upload a mask texture for KIND_MASKED_STAMP overlay primitives.
-    pub fn set_overlay_mask(
-        &mut self,
-        device: &wgpu::Device,
-        queue: &wgpu::Queue,
-        width: u32,
-        height: u32,
-        rgba: &[u8],
-    ) {
-        self.tool_overlay
-            .set_mask_texture(device, queue, width, height, rgba);
-        self.needs_present = true;
-    }
-
-    /// Drop the uploaded mask (fall back to 1×1 white).
-    pub fn clear_overlay_mask(&mut self) {
-        self.tool_overlay.clear_mask_texture();
-        self.needs_present = true;
-    }
-
-    /// Ensure the overlay's preview-mask texture exists at the given size;
-    /// returns a view for a brush node to render into.
-    pub fn ensure_overlay_preview_mask(
-        &mut self,
-        device: &wgpu::Device,
-        width: u32,
-        height: u32,
-    ) -> &wgpu::TextureView {
-        self.tool_overlay.ensure_preview_mask(device, width, height)
+    /// Mutable access to the tool overlay. Callers that change overlay state
+    /// must follow with `mark_needs_present()` themselves.
+    pub fn tool_overlay_mut(&mut self) -> &mut ToolOverlay {
+        &mut self.tool_overlay
     }
 
     /// Split-borrow accessor for the preview-render hot path: returns
@@ -1984,406 +1895,6 @@ impl Compositor {
         Option<&crate::gpu::selection::SelectionState>,
     ) {
         (&mut self.tool_overlay, self.selection_state.as_ref())
-    }
-
-    /// Route the preview-mask texture as the active overlay mask binding.
-    pub fn use_overlay_preview_mask(&mut self) {
-        self.tool_overlay.use_preview_mask_as_mask();
-        self.needs_present = true;
-    }
-
-    /// Unbind the preview mask (falls back to 1×1 white default).
-    pub fn clear_overlay_preview_mask(&mut self) {
-        self.tool_overlay.clear_preview_mask();
-        self.needs_present = true;
-    }
-
-    /// Borrow the overlay's preview-mask Texture (None if never allocated).
-    pub fn overlay_preview_mask_texture(&self) -> Option<&wgpu::Texture> {
-        self.tool_overlay.preview_mask_texture()
-    }
-
-    /// Immutable access to the tool overlay for test assertions.
-    pub fn tool_overlay_ref(&self) -> &ToolOverlay {
-        &self.tool_overlay
-    }
-
-    /// CPU-side hit test on overlay primitives.
-    pub fn overlay_hit_test(&self, screen_x: f32, screen_y: f32) -> Option<usize> {
-        self.tool_overlay.hit_test(screen_x, screen_y)
-    }
-
-    // --- Floating Content (Transform) ---
-    //
-    // The floating preview is a *derived view* of the target node's texture:
-    // when a transform is active, the compositor maintains a per-target
-    // preview texture rebuilt on every matrix update, holding "what the
-    // target would look like if commit ran right now". The render walk's
-    // `effective_node_view` and `effective_mask_bind_group` accessors swap
-    // the live view / mask bind group for the preview equivalents when the
-    // floating target is encountered, so the host's normal blend pass
-    // renders the preview without any extra render pass — and isolation,
-    // grouping, and other branch-free render concerns compose with it
-    // automatically.
-    //
-    // The compositor exposes primitives (set/clear floating content, update
-    // uniforms + preview, commit to live target). The engine drives them
-    // by calling `update_floating_preview` after each matrix change and on
-    // setup_transform.
-
-    /// Allocate the per-target preview texture (and, when target is R8, a
-    /// mask-shape bind group sampling it) plus the canvas-aligned blend
-    /// uniform buffer the host's blend pass reads when this layer is the
-    /// floating target.
-    ///
-    /// Preview is canvas-sized (not live-sized) so a translate that moves
-    /// content past the live layer's bounding box still has room on the
-    /// preview to write — clipped at canvas bounds, which is all the
-    /// viewport renders anyway. Commit's `grow_node_to_fit` separately
-    /// expands the live layer so off-canvas pixels survive the commit.
-    fn allocate_preview_resources(
-        &self,
-        device: &wgpu::Device,
-        format: wgpu::TextureFormat,
-    ) -> (
-        wgpu::Texture,
-        wgpu::TextureView,
-        Option<wgpu::BindGroup>,
-        wgpu::Buffer,
-    ) {
-        let preview = device.create_texture(&wgpu::TextureDescriptor {
-            label: Some("floating-preview"),
-            size: wgpu::Extent3d {
-                width: self.canvas_width.max(1),
-                height: self.canvas_height.max(1),
-                depth_or_array_layers: 1,
-            },
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format,
-            // COPY_SRC needed because `render_commit` runs
-            // `copy_for_compositing` against the render target before its
-            // shader pass — when the target is the preview texture, that
-            // copy reads from preview.
-            usage: wgpu::TextureUsages::TEXTURE_BINDING
-                | wgpu::TextureUsages::COPY_SRC
-                | wgpu::TextureUsages::COPY_DST
-                | wgpu::TextureUsages::RENDER_ATTACHMENT,
-            view_formats: &[],
-        });
-        let view = preview.create_view(&wgpu::TextureViewDescriptor::default());
-        let mask_bg = if format == wgpu::TextureFormat::R8Unorm {
-            Some(device.create_bind_group(&wgpu::BindGroupDescriptor {
-                label: Some("floating-preview-mask-bg"),
-                layout: &self.blend_pipelines.mask_bind_group_layout,
-                entries: &[wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: wgpu::BindingResource::TextureView(&view),
-                }],
-            }))
-        } else {
-            None
-        };
-        let blend_uniform_buf = device.create_buffer(&wgpu::BufferDescriptor {
-            label: Some("floating-preview-blend-uniforms"),
-            size: std::mem::size_of::<BlendUniforms>() as u64,
-            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-            mapped_at_creation: false,
-        });
-        (preview, view, mask_bg, blend_uniform_buf)
-    }
-
-    /// Look up the target's format + dimensions. Falls back to canvas-sized
-    /// RGBA8 when the node texture hasn't been allocated yet (paste-as-
-    /// floating creates the layer before its texture).
-    fn target_format_and_dims(&self, target_layer: LayerId) -> (wgpu::TextureFormat, u32, u32) {
-        match self.node_textures.get(&target_layer) {
-            Some(t) => {
-                let ext = t.layer_extent();
-                (t.format(), ext.width, ext.height)
-            }
-            None => (
-                wgpu::TextureFormat::Rgba8Unorm,
-                self.canvas_width,
-                self.canvas_height,
-            ),
-        }
-    }
-
-    /// Set up floating content for GPU preview from straight-alpha RGBA
-    /// pixel data (used by the paste paths). The target's preview texture
-    /// is allocated alongside.
-    pub fn set_floating_content(
-        &mut self,
-        device: &wgpu::Device,
-        queue: &wgpu::Queue,
-        rgba_data: &[u8],
-        _source_origin: (i32, i32),
-        source_width: u32,
-        source_height: u32,
-        target_layer: LayerId,
-    ) {
-        let (target_format, _tw, _th) = self.target_format_and_dims(target_layer);
-        let (preview_texture, preview_view, preview_mask_bg, preview_blend_uniform_buf) =
-            self.allocate_preview_resources(device, target_format);
-        self.transform_pass.set_floating_content(
-            device,
-            queue,
-            &self.sampler,
-            rgba_data,
-            source_width,
-            source_height,
-            target_layer,
-            target_format,
-            preview_texture,
-            preview_view,
-            preview_mask_bg,
-            preview_blend_uniform_buf,
-        );
-        // Seed the preview's blend uniforms from the live layer's cached
-        // props now that the floating session is active.
-        self.write_preview_blend_uniforms_if_active(queue, target_layer);
-        self.mark_dirty();
-    }
-
-    /// Set floating content by copying directly from a node's GPU texture.
-    /// GPU→GPU copy — no CPU tiles involved. Format dispatch (R8 mask vs RGBA
-    /// layer) is driven by the texture's own format from the unified node pool.
-    pub fn set_floating_content_from_gpu(
-        &mut self,
-        device: &wgpu::Device,
-        queue: &wgpu::Queue,
-        encoder: &mut wgpu::CommandEncoder,
-        source_origin: (i32, i32),
-        source_width: u32,
-        source_height: u32,
-        target_layer: LayerId,
-    ) {
-        let layer = match self.node_textures.get(&target_layer) {
-            Some(t) => t,
-            None => return,
-        };
-        let target_format = layer.format();
-        let (preview_texture, preview_view, preview_mask_bg, preview_blend_uniform_buf) =
-            self.allocate_preview_resources(device, target_format);
-        // Re-borrow `layer` after `allocate_preview_resources` — the helper
-        // doesn't take `&mut self`, but rust-analyzer prefers the explicit
-        // re-fetch over keeping the borrow live across the helper call.
-        let layer = self
-            .node_textures
-            .get(&target_layer)
-            .expect("layer present after preview allocation");
-        self.transform_pass.set_floating_content_from_gpu(
-            device,
-            queue,
-            encoder,
-            &self.sampler,
-            layer,
-            source_origin,
-            source_width,
-            source_height,
-            target_layer,
-            target_format,
-            preview_texture,
-            preview_view,
-            preview_mask_bg,
-            preview_blend_uniform_buf,
-        );
-        // Seed the preview's blend uniforms from the live layer's cached
-        // props now that the floating session is active.
-        self.write_preview_blend_uniforms_if_active(queue, target_layer);
-        self.mark_dirty();
-    }
-
-    /// Update the floating preview: write the current matrix uniforms, copy
-    /// live target → preview texture, apply the engine-side `clear_shape`
-    /// (None for paste mode, `Some` for transform mode), then run the
-    /// commit shader into the preview. The resulting preview texture is
-    /// what the host's blend pass reads through `effective_node_view` /
-    /// `effective_mask_bind_group` for the rest of the frame.
-    ///
-    /// Called by the engine on `setup_transform` (initial preview) and
-    /// `update_floating_matrix` (per drag tick).
-    #[allow(clippy::too_many_arguments)]
-    pub fn update_floating_preview(
-        &self,
-        device: &wgpu::Device,
-        queue: &wgpu::Queue,
-        paint_pipelines: &crate::gpu::paint_target::PaintPipelines,
-        matrix: &crate::gpu::transform::Affine2D,
-        source_origin: (i32, i32),
-        source_width: u32,
-        source_height: u32,
-        clear_shape: Option<&crate::gpu::transform::ClearShape>,
-    ) {
-        let Some(state) = self.transform_pass.active.as_ref() else {
-            return;
-        };
-        let live = match self.node_textures.get(&state.target_layer) {
-            Some(t) => t,
-            None => return,
-        };
-
-        // The preview is canvas-aligned: the transform shader writes the
-        // moved source content using target_offset=(0,0), target_size=canvas,
-        // so any pixel that lands within the canvas survives — including
-        // ones that fell outside the live texture's bounding box.
-        self.transform_pass.update_uniforms(
-            queue,
-            matrix,
-            source_origin,
-            source_width,
-            source_height,
-            (0, 0),
-            self.canvas_width,
-            self.canvas_height,
-            self.canvas_width,
-            self.canvas_height,
-        );
-
-        let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
-            label: Some("floating-preview-build"),
-        });
-
-        // 0. Reset the whole preview to transparent. The copy below only
-        //    repaints the canvas portion live actually covers, and the
-        //    commit shader discards outside the transformed source bounds —
-        //    so canvas pixels outside live's extent would otherwise retain
-        //    previous-frame transform writes (ghost trails).
-        crate::gpu::clear_view_transparent(
-            &mut encoder,
-            &state.preview_view,
-            "floating-preview-clear",
-        );
-
-        // 1. Copy live → preview so non-source-rect pixels are preserved.
-        //    Live texture sits at `(live.offset_x, live.offset_y)` in canvas
-        //    space; clip the copy to the on-canvas portion (the preview is
-        //    canvas-sized at origin 0,0). Off-canvas pixels are not in the
-        //    preview — the viewport never renders them anyway, and commit's
-        //    `grow_node_to_fit` preserves them on the live texture.
-        let canvas_rect =
-            crate::coord::CanvasRect::from_xywh(0, 0, self.canvas_width, self.canvas_height);
-        let live_canvas_extent = live.canvas_extent();
-        if let Some(visible) = live_canvas_extent.intersect(canvas_rect) {
-            // Translate the visible canvas slice into the live texture's
-            // layer-local coordinate frame for the GPU copy origin.
-            let visible_layer = live
-                .canvas_to_layer_rect(visible)
-                .expect("intersect with live's extent yields a layer-local rect");
-            let dst_x = visible.x0() as u32;
-            let dst_y = visible.y0() as u32;
-            encoder.copy_texture_to_texture(
-                wgpu::TexelCopyTextureInfo {
-                    texture: live.texture(),
-                    mip_level: 0,
-                    origin: wgpu::Origin3d {
-                        x: visible_layer.x0(),
-                        y: visible_layer.y0(),
-                        z: 0,
-                    },
-                    aspect: wgpu::TextureAspect::All,
-                },
-                wgpu::TexelCopyTextureInfo {
-                    texture: &state.preview_texture,
-                    mip_level: 0,
-                    origin: wgpu::Origin3d {
-                        x: dst_x,
-                        y: dst_y,
-                        z: 0,
-                    },
-                    aspect: wgpu::TextureAspect::All,
-                },
-                wgpu::Extent3d {
-                    width: visible.width,
-                    height: visible.height,
-                    depth_or_array_layers: 1,
-                },
-            );
-        }
-
-        // 2. Apply the source-rect clear (transform mode only — paste mode
-        //    leaves the preview as a copy of the live target so the commit
-        //    shader composites over the existing pixels). The preview is
-        //    canvas-aligned, so the paint target reports canvas dims/offset.
-        if let Some(cs) = clear_shape {
-            let preview_target = crate::gpu::paint_target::GpuPaintTarget::from_canvas_texture(
-                &state.preview_texture,
-                &state.preview_view,
-                state.target_format,
-                self.canvas_width,
-                self.canvas_height,
-            );
-            match cs {
-                crate::gpu::transform::ClearShape::Rect(rect) => {
-                    preview_target.clear_rect(&mut encoder, paint_pipelines, queue, *rect);
-                }
-                crate::gpu::transform::ClearShape::Selection { mask_bind_group } => {
-                    preview_target.erase_with_selection(
-                        &mut encoder,
-                        paint_pipelines,
-                        queue,
-                        mask_bind_group,
-                    );
-                }
-            }
-        }
-
-        // 3. Run the commit shader into the preview at the current matrix.
-        self.transform_pass.render_commit(
-            device,
-            &mut encoder,
-            &state.preview_texture,
-            &state.preview_view,
-        );
-
-        queue.submit(std::iter::once(encoder.finish()));
-    }
-
-    /// Render the transform directly into the live target texture.
-    /// Used by `commit_floating()` after the engine has applied the
-    /// `clear_shape` to the live target.
-    pub fn commit_floating_to_texture(
-        &mut self,
-        device: &wgpu::Device,
-        encoder: &mut wgpu::CommandEncoder,
-        queue: &wgpu::Queue,
-        matrix: &crate::gpu::transform::Affine2D,
-        source_origin: (i32, i32),
-        source_width: u32,
-        source_height: u32,
-    ) {
-        let Some(state) = self.transform_pass.active.as_ref() else {
-            return;
-        };
-        let live = match self.node_textures.get(&state.target_layer) {
-            Some(t) => t,
-            None => return,
-        };
-
-        let live_extent = live.canvas_extent();
-        self.transform_pass.update_uniforms(
-            queue,
-            matrix,
-            source_origin,
-            source_width,
-            source_height,
-            (live_extent.x0(), live_extent.y0()),
-            live_extent.width,
-            live_extent.height,
-            self.canvas_width,
-            self.canvas_height,
-        );
-
-        self.transform_pass
-            .render_commit(device, encoder, live.texture(), live.view());
-    }
-
-    /// Remove floating content GPU state.
-    pub fn clear_floating_content(&mut self) {
-        self.transform_pass.clear();
-        self.mark_dirty();
     }
 
     /// Effective mask bind group for a host raster/group during compositing
@@ -2429,20 +1940,6 @@ impl Compositor {
             }
         }
         live_or_default
-    }
-
-    /// Get a reference to the transform source texture and its view.
-    /// Returns None if no floating content is active.
-    pub fn transform_source_texture(&self) -> Option<(&wgpu::Texture, &wgpu::TextureView)> {
-        self.transform_pass
-            .active
-            .as_ref()
-            .map(|s| (&s.source_texture, &s.source_view))
-    }
-
-    /// Check if floating content is active.
-    pub fn has_floating_content(&self) -> bool {
-        self.transform_pass.active.is_some()
     }
 
     /// Run the present pass, veil chain, and final blit to surface.
@@ -3253,37 +2750,8 @@ impl Compositor {
         self.veil_chain.clear_needs_present();
     }
 
-    /// Composite layers if needed, then present to an arbitrary texture view.
-    ///
-    /// This is the backend-agnostic rendering entry point. Any frontend
-    /// (WASM surface, native window, CEF hole-punch, headless test) can
-    /// provide a `TextureView` and get the composited + veiled result.
-    pub fn render_to_view(
-        &mut self,
-        device: &wgpu::Device,
-        queue: &wgpu::Queue,
-        target_view: &wgpu::TextureView,
-        doc: &mut Document,
-    ) {
-        if !self.has_pending_work(doc) {
-            return;
-        }
-
-        self.render_offscreen(device, queue, doc);
-
-        let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
-            label: Some("present-to-view"),
-        });
-        self.present_and_veils(&mut encoder, target_view);
-        queue.submit(std::iter::once(encoder.finish()));
-
-        self.finish_present();
-    }
-
     /// Upload dirty tiles, composite changed layers, present to a surface.
-    ///
-    /// Convenience wrapper around `render_to_view` that handles surface
-    /// acquisition and presentation. Used by the WASM frontend.
+    /// Used by the WASM frontend.
     pub fn render(
         &mut self,
         device: &wgpu::Device,
